+++ chef-server/.chef/knife.rb
@@ -1,0 +1,8 @@
+log_level                :info
+log_location             STDOUT
+node_name                'zvickery'
+client_key               '/Users/zvickery/dev/bsu-cloud-computing/chef-server/.chef/zvickery.pem'
+validation_client_name   'validation.pem'
+validation_key           '/Users/zvickery/dev/bsu-cloud-computing/chef-server/.chef/validation.pem'
+chef_server_url          'https://chef.zdvickery.us:443/organizations/zdv'
+syntax_check_cache_path  '/Users/zvickery/dev/bsu-cloud-computing/chef-server/.chef/syntax_check_cache'